requirements/base.txt
flask
Frozen-Flask
python-dotenv
